requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
dnspython==2.2.1
idna==3.3
requests==2.28.1
urllib3==1.26.12
